pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kdream"
version = "0.1.0"
description = "A Python library for orchestrating AI model workflows and recipes across multiple backends."
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "ismael faro" },
]
keywords = [
    "ai",
    "machine-learning",
    "generative-ai",
    "llm",
    "image-generation",
    "text-generation",
    "audio-generation",
    "video-generation",
    "recipes",
    "workflows",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "uv",
    "huggingface_hub",
    "pyyaml",
    "click",
    "pydantic>=2",
    "rich",
    "httpx",
    "gitpython",
    "anthropic",
]

[project.optional-dependencies]
runpod = ["runpod"]
kubernetes = ["kubernetes"]

[project.scripts]
kdream = "kdream.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["kdream"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
